+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 
-- Added greentea tests.
+- Added greentea tests
+- FIFO mode for and accelerometer
 
 ## [0.1.0] - 2018-07-29
 ### Added